Begin converting to library, update weather-now
The weather-now utility (and future vane utilities) should rely on
vane as a library/API, and serve as (hopefully useful!) examples
of how you can use vane.

This commit introduces the first pieces of the vane library. The
weather-now utility now uses the vane library to fetch current
weather conditions.

For now, errors are passed up in an API style (when possible)
to whatever is consuming the requested data. This will probably be
revisited later.