Support Weather Underground API
This is the first implementation of the Weather Underground API
within the vane library. This introduces the provider and API key
options in the weather-now utility.

The goal is to provide a consistent interface to third-party APIs
while handling each provider's quirks and gracefully degrading if
certain metrics aren't available across all APIs.

For now, the API differences between OpenWeatherMap and Weather
Underground mean that results may be somewhat inconsistent or
unexpected (i.e.: handling of units). This will be resolved in
future iterations, even to the extent of doing internal conversions
if necessary.